backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0